requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
xhtml2pdf==0.2.11
reportlab>=4.1.0
svglib==1.5.1
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.0
agno>=2.5.0
openai>=1.0.0
google-search-results>=2.4.0
